CAN actuator works with flight configuration. Wierd bug where the actuator will flood can with requests even though timeouts have been implemented

+++ src/RocketComponents/packets/nrcpackets.h
@@ -55,7 +55,7 @@ class NRCStatePacket : public RnpPacket{
          */
         void serialize(std::vector<uint8_t>& buf) override;
 
-        uint8_t state;
+        uint16_t state;
         int32_t value;
 
         static constexpr size_t size(){
